Improve subquery, including many bugs fixed.
