+++ src/Interpreters/IInterpreterUnionOrSelectQuery.cpp
@@ -4,6 +4,7 @@
 #include <Processors/QueryPlan/QueryPlan.h>
 #include <Processors/QueryPlan/BuildQueryPipelineSettings.h>
 #include <Processors/QueryPlan/Optimizations/QueryPlanOptimizationSettings.h>
+#include <Processors/QueryPlan/DistributedPlanner.h>
 
 namespace DB
 {
@@ -13,6 +14,16 @@ void IInterpreterUnionOrSelectQuery::extendQueryLogElemImpl(QueryLogElement & el
     elem.query_kind = "Select";
 }
 
+void IInterpreterUnionOrSelectQuery::rewriteDistributedQuery(bool is_subquery, [[maybe_unused]] size_t tables_count)
+{
+    /// Reset these fields to not share them with parent query.
+    if (is_subquery)
+    {
+        getContext()->resetQueryPlanFragmentInfo();
+        getContext()->getClientInfo().query_kind = ClientInfo::QueryKind::INITIAL_QUERY;
+        getContext()->getClientInfo().current_query_id = getContext()->generateQueryId();
+    }
+}
 
 QueryPipelineBuilder IInterpreterUnionOrSelectQuery::buildQueryPipeline()
 {
@@ -20,10 +31,13 @@ QueryPipelineBuilder IInterpreterUnionOrSelectQuery::buildQueryPipeline()
 
     buildQueryPlan(query_plan);
 
-    context->setSelectQuery(queryToString(this->query_ptr));
-    bool is_built = query_plan.buildDistributedPlan(context);
+    query_plan.checkInitialized();
+    query_plan.optimize(QueryPlanOptimizationSettings::fromContext(context));
 
-    QueryPlanOptimizationSettings optimization_settings{.optimize_plan = !is_built};
+    DistributedPlanner planner(query_plan, context);
+    planner.buildDistributedPlan();
+
+    QueryPlanOptimizationSettings optimization_settings{.optimize_plan = false};
     return std::move(*query_plan.buildQueryPipeline(
         optimization_settings, BuildQueryPipelineSettings::fromContext(context)));
 }
