+++ src/Processors/QueryPlan/QueryPlan.h
@@ -37,6 +37,8 @@ class LimitStep;
 struct QueryPlanOptimizationSettings;
 struct BuildQueryPipelineSettings;
 
+class DistirbutedPlanner;
+
 namespace JSONBuilder
 {
     class IItem;
@@ -65,6 +67,7 @@ using InterpreterParamsPtr = std::shared_ptr<InterpreterParams>;
 class QueryPlan
 {
 public:
+    friend class DistirbutedPlanner;
     QueryPlan();
     ~QueryPlan();
     QueryPlan(QueryPlan &&);
@@ -81,12 +84,7 @@ public:
 
     void reset();
 
-    void buildStages(ContextPtr context);       /// Used by initial node.
-    void debugStages();
-    /// Return true if result stage is moved forward.
-    bool scheduleStages(ContextMutablePtr context);    /// Used by initial node.
-    void buildPlanFragment(ContextPtr context); /// Used by non-initial nodes.
-    bool buildDistributedPlan(ContextMutablePtr context);
+    void collectCreatingSetPlan(std::vector<std::unique_ptr<QueryPlan>> & creating_set_plans);
 
     QueryPipelineBuilderPtr buildQueryPipeline(
         const QueryPlanOptimizationSettings & optimization_settings,
@@ -140,49 +138,7 @@ public:
 
     using Nodes = std::list<Node>;
 
-    struct Stage
-    {
-        int id; /// Current stage id.
-        std::vector<Stage *> parents = {}; /// Previous stages that current stage directly depends on.
-        Stage * child = nullptr;
-        std::vector<std::shared_ptr<String>> workers; /// Replicas that current stage should be executed on.
-        std::vector<std::shared_ptr<String>> sinks; /// Child's workers.
-        Node * root_node; /// Current stage's root node.
-        std::vector<Node *> leaf_nodes; /// Store leaf nodes which are from right side to left side.
-        bool is_leaf_stage = false; /// Current stage is a leaf stage if it has any leaf node reading data from storage(not from remote).
-        bool maybe_has_view_source = false; /// Current stage reads data to trigger materialized view.
-        bool has_input_function = false;
-    };
-
-    /// Note: do not use vector, otherwise pointers to elements in it will be invalidated when vector increases.
-    using Stages = std::list<Stage>;
-
-    struct PlanFragmentInfo
-    {
-        PlanFragmentInfo(int stage_id_, const String & node_id_, const std::vector<String> & sources_, const std::vector<String> & sinks_)
-            : stage_id(stage_id_), node_id(node_id_), sources(sources_), sinks(sinks_) {}
-        int stage_id;
-        String node_id; /// The replica name of plan fragment receiver, used by DistributedSource.
-        std::vector<String> sources; /// Point to the nodes sending data.
-        std::vector<String> sinks; /// Point to the nodes receiving data.
-    };
-    using PlanFragmentInfoPtr = std::shared_ptr<PlanFragmentInfo>;
-
-    struct CheckShuffleResult
-    {
-        bool is_shuffle = false;
-        UnionStep * current_union_step = nullptr;
-        JoinStep * current_join_step = nullptr;
-        AggregatingStep * child_aggregating_step = nullptr;
-        SortingStep * child_sorting_step = nullptr;
-        LimitStep * current_limit_step = nullptr;
-        LimitStep * child_limit_step = nullptr;
-    };
-    void checkShuffle(Node * current_node, Node * child_node, CheckShuffleResult & result);
-    String debugLocalPlanFragment(const String & query_id, int stage_id, const String & node_id, const std::vector<Node *> distributed_source_nodes);
-    String debugRemotePlanFragment(const String & query, const String & receiver, const String & query_id, const Stage * stage);
-
-private:
+public:
     Nodes nodes;
     Node * root = nullptr;
 
@@ -193,21 +149,6 @@ private:
     size_t max_threads = 0;
     std::vector<ContextPtr> interpreter_context;
 
-    Stages stages;
-    Stage * result_stage = nullptr;
-
-    struct pairHasher {
-        template <class T1, class T2>
-        size_t operator()(const std::pair<T1, T2> & p) const
-        {
-            auto hash1 = std::hash<T1>{}(p.first);
-            auto hash2 = std::hash<T2>{}(p.second);
-            return hash1 ^ hash2;
-        }
-    };
-
-    /// Key is {stage_id, receiver_address}.
-    std::unordered_map<std::pair<int, std::shared_ptr<String>>, PlanFragmentInfoPtr, pairHasher> plan_fragment_infos;
     Poco::Logger * log;
 };
